fix: moved test utils from utils
