+++ plugin/src/shared/utils/index.ts
@@ -6,4 +6,3 @@ export * from './config-utils';
 export * from './ts-utils';
 export * from './constants';
 export * from './get-lib-directory-name';
-export * from './tests-utils';
